% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/righttriangle_third_side.R
\name{righttriangle_third_side}
\alias{righttriangle_third_side}
\title{calculates the length of the third side of right triangle}
\usage{
righttriangle_third_side(a = 0, b = 0, c = 0)
}
\arguments{
\item{a}{numeric value-}

\item{b}{numeric value-}

\item{c}{numeric value-}
}
\value{
a number
}
\description{
The Pythagorean theorem states that the square of the hypotenuse (the side opposite the right angle) is equal to the sum of the squares of the other two side. Given the lengths of two sides of the triangle, this function calculates the length of the third side.
}
\examples{
righttriangle_third_side(a = 3, b = 4, ) # Calculate c
righttriangle_third_side(b = 4, c = 5, ) # Calculate a
righttriangle_third_side(a = 3, c = 5, ) # Calculate b
}
