% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_fun.R
\name{trim_fun}
\alias{trim_fun}
\title{Trimmed mean function}
\usage{
trim_fun(x, s, l)
}
\arguments{
\item{x}{A numeric value-}
}
\value{
a number
}
\description{
trimmed mean function that calculates the mean of a numeric vector, ignoring the smallest and largest values
}
\examples{
x<-c(2,3,0,6,1,22,1,6,88)
s<-2
l<-2
trim_fun(x,s,l)
}
